refactor effective rate calculation
